Struct enr::Enr

pub struct Enr<K: EnrKey> { /* private fields */ }

The ENR, allowing for arbitrary signing algorithms.

This struct will always have a valid signature, known public key type, sequence number and NodeId. All other parameters are variable/optional.

Implementations

impl<K: EnrKey> Enr<K>

pub const fn node_id(&self) -> NodeId

The NodeId for the record.

pub const fn seq(&self) -> u64

The current sequence number of the ENR record.

pub fn get(&self, key: impl AsRef<[u8]>) -> Option<&[u8]>

Reads a custom key from the record if it exists, decoded as data.

pub fn get_decodable<T: Decodable>( &self, key: impl AsRef<[u8]> ) -> Option<Result<T, DecoderError>>

Reads a custom key from the record if it exists, decoded as T.

pub fn get_raw_rlp(&self, key: impl AsRef<[u8]>) -> Option<&[u8]>

Reads a custom key from the record if it exists as raw RLP bytes.

pub fn iter(&self) -> impl Iterator<Item = (&Vec<u8>, &[u8])>

Returns an iterator over all key/value pairs in the ENR.

pub fn ip4(&self) -> Option<Ipv4Addr>

Returns the IPv4 address of the ENR record if it is defined.

pub fn ip6(&self) -> Option<Ipv6Addr>

Returns the IPv6 address of the ENR record if it is defined.

pub fn id(&self) -> Option<String>

The id of ENR record if it is defined.

pub fn tcp4(&self) -> Option<u16>

The TCP port of ENR record if it is defined.

pub fn tcp6(&self) -> Option<u16>

The IPv6-specific TCP port of ENR record if it is defined.

pub fn udp4(&self) -> Option<u16>

The UDP port of ENR record if it is defined.

pub fn udp6(&self) -> Option<u16>

The IPv6-specific UDP port of ENR record if it is defined.

pub fn udp4_socket(&self) -> Option<SocketAddrV4>

Provides a socket (based on the UDP port), if the IPv4 and UDP fields are specified.

pub fn udp6_socket(&self) -> Option<SocketAddrV6>

Provides a socket (based on the UDP port), if the IPv6 and UDP fields are specified.

pub fn tcp4_socket(&self) -> Option<SocketAddrV4>

Provides a socket (based on the TCP port), if the IP and TCP fields are specified.

pub fn tcp6_socket(&self) -> Option<SocketAddrV6>

Provides a socket (based on the TCP port), if the IPv6 and TCP6 fields are specified.

pub fn signature(&self) -> &[u8]

The signature of the ENR record.

pub fn public_key(&self) -> K::PublicKey

Returns the public key of the ENR record.

Panics

Will panic if the public key is not supported.

pub fn verify(&self) -> bool

Verify the signature of the ENR record.

pub fn to_base64(&self) -> String

Provides the URL-safe base64 encoded “text” version of the ENR prefixed by “enr:”.

pub fn size(&self) -> usize

Returns the current size of the ENR.

pub fn set_seq(&mut self, seq: u64, key: &K) -> Result<(), EnrError>

Allows setting the sequence number to an arbitrary value.

pub fn insert<T: Encodable>( &mut self, key: impl AsRef<[u8]>, value: &T, enr_key: &K ) -> Result<Option<Bytes>, EnrError>

Adds or modifies a key/value to the ENR record. A EnrKey is required to re-sign the record once modified.

Returns the previous value as rlp encoded bytes in the record if it exists.

pub fn insert_raw_rlp( &mut self, key: impl AsRef<[u8]>, value: Bytes, enr_key: &K ) -> Result<Option<Bytes>, EnrError>

Adds or modifies a key/value to the ENR record. A EnrKey is required to re-sign the record once modified. The value here is interpreted as raw RLP data.

Returns the previous value as rlp encoded bytes in the record if it exists.

pub fn set_ip( &mut self, ip: IpAddr, key: &K ) -> Result<Option<IpAddr>, EnrError>

Sets the ip field of the ENR. Returns any pre-existing IP address in the record.

pub fn set_udp4(&mut self, udp: u16, key: &K) -> Result<Option<u16>, EnrError>

Sets the udp field of the ENR. Returns any pre-existing UDP port in the record.

pub fn set_udp6(&mut self, udp: u16, key: &K) -> Result<Option<u16>, EnrError>

Sets the udp6 field of the ENR. Returns any pre-existing UDP port in the record.

pub fn set_tcp4(&mut self, tcp: u16, key: &K) -> Result<Option<u16>, EnrError>

Sets the tcp field of the ENR. Returns any pre-existing tcp port in the record.

pub fn set_tcp6(&mut self, tcp: u16, key: &K) -> Result<Option<u16>, EnrError>

Sets the tcp6 field of the ENR. Returns any pre-existing tcp6 port in the record.

pub fn set_udp_socket( &mut self, socket: SocketAddr, key: &K ) -> Result<(), EnrError>

Sets the IP and UDP port in a single update with a single increment in sequence number.

pub fn set_tcp_socket( &mut self, socket: SocketAddr, key: &K ) -> Result<(), EnrError>

Sets the IP and TCP port in a single update with a single increment in sequence number.

pub fn remove_insert<'a>( &mut self, remove_keys: impl Iterator<Item = impl AsRef<[u8]>>, insert_key_values: impl Iterator<Item = (impl AsRef<[u8]>, &'a [u8])>, enr_key: &K ) -> Result<(Vec<Option<Bytes>>, Vec<Option<Bytes>>), EnrError>

Removes key/value mappings and adds or overwrites key/value mappings to the ENR record as one sequence number update. An EnrKey is required to re-sign the record once modified. Reverts whole ENR record on error.

Returns the previous values as rlp encoded bytes if they exist for the removed and added/ overwritten keys.

pub fn set_public_key( &mut self, public_key: &K::PublicKey, key: &K ) -> Result<(), EnrError>

Sets a new public key for the record.

pub fn is_udp_reachable(&self) -> bool

Returns wether the node can be reached over UDP or not.

pub fn is_tcp_reachable(&self) -> bool

Returns wether the node can be reached over TCP or not.