Trait enr::EnrKey

pub trait EnrKey: Send + Sync + Unpin + 'static {
    type PublicKey: EnrPublicKey + Clone;

    // Required methods
    fn sign_v4(&self, msg: &[u8]) -> Result<Vec<u8>, SigningError>;
    fn public(&self) -> Self::PublicKey;
    fn enr_to_public(
        content: &BTreeMap<Vec<u8>, Bytes>
    ) -> Result<Self::PublicKey, DecoderError>;
}

The trait required for a key to sign and modify an ENR record.

Required Associated Types

type PublicKey: EnrPublicKey + Clone

Required Methods

fn sign_v4(&self, msg: &[u8]) -> Result<Vec<u8>, SigningError>

Performs ENR-specific signing for the v4 identity scheme.

fn public(&self) -> Self::PublicKey

Returns the public key associated with current key pair.

fn enr_to_public( content: &BTreeMap<Vec<u8>, Bytes> ) -> Result<Self::PublicKey, DecoderError>

Provides a method to decode a raw public key from an ENR BTreeMap to a useable public key.

This method allows a key type to decode the raw bytes in an ENR to a useable EnrPublicKey. It takes the ENR’s BTreeMap and returns a public key.

Note: This specifies the supported key schemes for an ENR.

Object Safety

This trait is not object safe.